feat: Discussion 도메인 Repository 정의 및 DTO 수정

+++ Dialog/core/network/src/commonMain/kotlin/com/on/network/dto/request/DiscussionSummaryRequest.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DiscussionSummaryRequest(
     @SerialName("discussionId")
-    val discussionId: Int,
+    val discussionId: Long,
 )
